Admin To Store IPFS

+++ utils/utils.js
@@ -1,5 +1,5 @@
 const rimraf = require("rimraf");
-var fs = require('fs');
+var fs = require("fs");
 const ipfsClient = require("ipfs-http-client");
 const ganache = require("ganache-cli");
 const mongoose = require("mongoose");
@@ -8,9 +8,10 @@ const Web3 = require("web3");
 const User = require("../models/user");
 const Document = require("../models/document");
 
-exports.getPendingDocuments = async (department, currentPage, perPage) => {
-  return Document.find({ depId: department, status: "pending" })
+exports.getDocuments = async (department, currentPage, perPage, status) => {
+  return Document.find({ depId: department, status: status })
     .populate("user", {
+      name: 1,
       dob: 1,
       fathersName: 1,
       mothersName: 1,
@@ -21,33 +22,14 @@ exports.getPendingDocuments = async (department, currentPage, perPage) => {
     .limit(perPage);
 };
 
-exports.getVerifiedDocuments = async (department, currentPage, perPage) => {
-  return Document.find({ depId: department, status: "verified" })
-    .populate("user", {
-      dob: 1,
-      fathersName: 1,
-      mothersName: 1,
-      address: 1,
-      phone: 1,
-    })
-    .skip((currentPage - 1) * perPage)
-    .limit(perPage);
-};
-
-exports.getFailedDocuments = async (department, currentPage, perPage) => {
-  return Document.find({ depId: department, status: "verified" })
-    .populate("user", {
-      dob: 1,
-      fathersName: 1,
-      mothersName: 1,
-      address: 1,
-      phone: 1,
-    })
-    .skip((currentPage - 1) * perPage)
-    .limit(perPage);
-};
-
-exports.startVerification = async (department, status, file, res, next) => {
+exports.startVerification = async (
+  department,
+  uid,
+  status,
+  file,
+  res,
+  next
+) => {
   if (!file) {
     if (status == "verified") {
       const error = new Error("No image provided.");
@@ -57,12 +39,12 @@ exports.startVerification = async (department, status, file, res, next) => {
   }
 
   const doc = await Document.findOne({ depId: department });
-
+  console.log(department);
   Document.updateOne(
-    { depId: department },
+    { user: uid, depId: department },
     {
       status: status,
-      file: status=="fail" ? doc.file : file.path,
+      file: status == "fail" ? doc.file : file.path,
     },
     function (err, raw) {
       if (err) {
@@ -70,7 +52,8 @@ exports.startVerification = async (department, status, file, res, next) => {
           err.statusCode = 500;
         }
         next(err);
-      } else {
+      }
+      if (raw.nModified > 0) {
         if (status == "verified") {
           fs.rmSync(doc.file, {
             force: true,
@@ -81,6 +64,13 @@ exports.startVerification = async (department, status, file, res, next) => {
         res.status(200).json({
           message: "Verified successfully.",
         });
+      } else {
+        if (err) {
+          if (!err.statusCode) {
+            err.statusCode = 500;
+          }
+          next(err);
+        }
       }
     }
   );
